Cast parameters of mktime to int for php 8

+++ Configuration/TCA/tx_styleguide_elements_basic.php
@@ -633,7 +633,7 @@ return [
                 'eval' => 'datetime',
                 'default' => 0,
                 'range' => [
-                    'lower' => mktime(0, 0, 0, date('m'), date('d'), date('Y'))
+                    'lower' => mktime(0, 0, 0, (int)date('m'), (int)date('d'), (int)date('Y'))
                 ]
             ],
         ],
